Add layer + layerstack and add glm + glad

+++ HideAndSeek/src/Engine/Event/MouseEvent.h
@@ -52,8 +52,8 @@ namespace Engine {
 		MouseScrolledEvent(double xOffset, double yOffset) 
 			: m_XOffset(xOffset), m_YOffset(yOffset) {}
 
-		inline float GetXOffset() const { return m_XOffset; }
-		inline float GetYOffset() const { return m_YOffset; }
+		inline double GetXOffset() const { return m_XOffset; }
+		inline double GetYOffset() const { return m_YOffset; }
 
 		std::string ToString() const override {
 			std::stringstream ss;
@@ -74,8 +74,8 @@ namespace Engine {
 		MouseMovedEvent(double x, double y)
 			: m_MouseX(x), m_MouseY(y) {}
 
-		inline float GetX() const { return m_MouseX; }
-		inline float GetY() const { return m_MouseY; }
+		inline double GetX() const { return m_MouseX; }
+		inline double GetY() const { return m_MouseY; }
 
 		std::string ToString() const override {
 			std::stringstream ss;
